Add auth code
