+++ API/CCW.Admin/Program.cs
@@ -5,7 +5,9 @@ using CCW.Admin.Entities;
 using CCW.Admin.Mappers;
 using CCW.Admin.Models;
 using CCW.Admin.Services;
+using CCW.Common.AuthorizationPolicies;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.IdentityModel.Tokens;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -19,6 +21,10 @@ builder.Services.AddSingleton<ICosmosDbService>(
 
 builder.Services.AddSingleton<IMapper<AgencyProfileSettingsRequestModel, AgencyProfileSettings>, AgencyProfileRequestSettingsModelToEntityMapper>();
 builder.Services.AddSingleton<IMapper<AgencyProfileSettings, AgencyProfileSettingsResponseModel>, EntityToAgencyProfileSettingsResponseModelMapper>();
+
+builder.Services.AddScoped<IAuthorizationHandler, IsAdminHandler>();
+builder.Services.AddScoped<IAuthorizationHandler, IsSystemAdminHandler>();
+builder.Services.AddScoped<IAuthorizationHandler, IsProcessorHandler>();
 
 builder.Services
     .AddAuthentication("aad")
@@ -47,6 +53,36 @@ builder.Services
         {
             OnAuthenticationFailed = AuthenticationFailed,
         };
+    });
+
+builder.Services
+    .AddAuthorization(options =>
+    {
+        var apiPolicy = new AuthorizationPolicyBuilder("aad", "b2c")
+            .AddAuthenticationSchemes("aad", "b2c")
+            .RequireAuthenticatedUser()
+            .Build();
+
+        options.AddPolicy("ApiPolicy", apiPolicy);
+
+        options.AddPolicy("RequireAdminOnly",
+            policy =>
+            {
+                policy.RequireRole("CCW-ADMIN-ROLE");
+                policy.Requirements.Add(new RoleRequirement("CCW-ADMIN-ROLE"));
+            });
+
+        options.AddPolicy("RequireSystemAdminOnly", policy =>
+        {
+            policy.RequireRole("CCW-SYSTEM-ADMINS-ROLE");
+            policy.Requirements.Add(new RoleRequirement("CCW-SYSTEM-ADMINS-ROLE"));
+        });
+
+        options.AddPolicy("RequireProcessorOnly", policy =>
+        {
+            policy.RequireRole("CCW-PROCESSORS-ROLE");
+            policy.Requirements.Add(new RoleRequirement("CCW-PROCESSORS-ROLE"));
+        });
     });
 
 builder.Services.AddControllers();
