+++ API/CCW.Application/Program.cs
@@ -5,7 +5,9 @@ using CCW.Application.Entities;
 using CCW.Application.Mappers;
 using CCW.Application.Models;
 using CCW.Application.Services;
+using CCW.Common.AuthorizationPolicies;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.IdentityModel.Tokens;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -64,6 +66,10 @@ builder.Services.AddSingleton<IMapper<PermitApplicationRequestModel, SpouseAddre
 builder.Services.AddSingleton<IMapper<PermitApplicationRequestModel, UploadedDocument[]>, RequestPermitApplicationToUploadDocumentMapper>();
 builder.Services.AddSingleton<IMapper<PermitApplicationRequestModel, BackgroudCheck>, RequestPermitApplicationToBackgroundCheckMapper>();
 
+builder.Services.AddScoped<IAuthorizationHandler, IsAdminHandler>();
+builder.Services.AddScoped<IAuthorizationHandler, IsSystemAdminHandler>();
+builder.Services.AddScoped<IAuthorizationHandler, IsProcessorHandler>();
+
 builder.Services
     .AddAuthentication("aad")
     .AddJwtBearer("aad", o =>
@@ -91,6 +97,36 @@ builder.Services
         {
             OnAuthenticationFailed = AuthenticationFailed,
         };
+    });
+
+builder.Services
+    .AddAuthorization(options =>
+    {
+        var apiPolicy = new AuthorizationPolicyBuilder("aad", "b2c")
+            .AddAuthenticationSchemes("aad", "b2c")
+            .RequireAuthenticatedUser()
+            .Build();
+
+        options.AddPolicy("ApiPolicy", apiPolicy);
+
+        options.AddPolicy("RequireAdminOnly",
+            policy =>
+            {
+                policy.RequireRole("CCW-ADMIN-ROLE");
+                policy.Requirements.Add(new RoleRequirement("CCW-ADMIN-ROLE"));
+            });
+
+        options.AddPolicy("RequireSystemAdminOnly", policy =>
+        {
+            policy.RequireRole("CCW-SYSTEM-ADMINS-ROLE");
+            policy.Requirements.Add(new RoleRequirement("CCW-SYSTEM-ADMINS-ROLE"));
+        });
+
+        options.AddPolicy("RequireProcessorOnly", policy =>
+        {
+            policy.RequireRole("CCW-PROCESSORS-ROLE");
+            policy.Requirements.Add(new RoleRequirement("CCW-PROCESSORS-ROLE"));
+        });
     });
 
 builder.Services.AddControllers();
