+++ API/CCW.Common/CCW.Common.csproj
@@ -5,4 +5,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="7.0.0" />
+  </ItemGroup>
+
 </Project>
