+++ API/CCW.Payment/CCW.Payment.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CCW.Common\CCW.Common.csproj" />
+  </ItemGroup>
+
 </Project>
