+++ API/CCW.Schedule/CCW.Schedule.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CCW.Common\CCW.Common.csproj" />
+  </ItemGroup>
+
 </Project>
