--- API/CCW.UserProfile/AuthorizationPolicies/RoleRequirement.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-
-namespace CCW.UserProfile.AuthorizationPolicies;
-
-public class RoleRequirement : IAuthorizationRequirement
-{
-    protected string Role { get; set; }
-
-    public RoleRequirement(string role)
-    {
-        Role = role;
-    }
-}
