+++ API/CCW.UserProfile/CCW.UserProfile.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CCW.Common\CCW.Common.csproj" />
+  </ItemGroup>
+
 </Project>
