+++ API/CCW.sln
@@ -27,6 +27,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCW.Schedule.Tests", "CCW.S
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCW.UserProfile.Tests", "CCW.UserProfile.Tests\CCW.UserProfile.Tests.csproj", "{257E179F-8953-426B-BF63-44BA33F3ADAD}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCW.Common", "CCW.Common\CCW.Common.csproj", "{C5BAE9D6-BD64-4AF9-9FF0-B74CD47D5539}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -81,6 +83,10 @@ Global
 		{257E179F-8953-426B-BF63-44BA33F3ADAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{257E179F-8953-426B-BF63-44BA33F3ADAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{257E179F-8953-426B-BF63-44BA33F3ADAD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5BAE9D6-BD64-4AF9-9FF0-B74CD47D5539}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5BAE9D6-BD64-4AF9-9FF0-B74CD47D5539}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5BAE9D6-BD64-4AF9-9FF0-B74CD47D5539}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5BAE9D6-BD64-4AF9-9FF0-B74CD47D5539}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
